chapter 1

+++ chapter-1/1-3.c
@@ -5,7 +5,7 @@
 */  
 void main()
 {
-    int fahrenheit, celsius;
+    float fahrenheit, celsius;
     int lower, upper, step;
 
     lower = 0;   // lowest temp
@@ -14,7 +14,7 @@ void main()
     fahrenheit = lower;
     while (fahrenheit <= upper) {
         celsius = 5 * (fahrenheit - 32) / 9;
-        printf("%3d\t%6d\n", fahrenheit, celsius);
+        printf("%3.0f\t%6.1f\n", fahrenheit, celsius);
         fahrenheit = fahrenheit + step;
     }
 }
